requirements.txt
opencv-python
numpy
Pillow
tkinter
ttkthemes
